+++ inc/vplfile.h
@@ -7,7 +7,7 @@ struct vpl_header
 {
 	uint32_t remap_start;
 	uint32_t remap_end;
-	size_t section_count;
+	uint32_t section_count;
 	uint32_t _reserved;
 };
